extract woektime widget to file
